# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: TrustMate API
  version: "1"
  summary: Self-contained Certificate Authority with CRL, OCSP, RFC 3161 TSA and ACME.
  description: |
    TrustMate issues X.509 certificates from a root/intermediate CA it
    bootstraps itself, and optionally serves revocation (CRL, OCSP), RFC 3161
    timestamps, and RFC 8555 ACME enrollment.

    Authentication is mutual TLS. Operations marked with the `mtls` security
    requirement need a client certificate issued by this CA that has a role
    assigned: `manager` (certificate operations) or `admin` (everything,
    including client onboarding and configuration). The required role is
    stated in each operation's `x-required-role`. Public PKI endpoints (CA
    certificates, CRL, OCSP, TSA) and the ACME client-facing routes need no
    client certificate; ACME requests authenticate with their own JWS
    signatures instead.

    Route groups tagged `revocation`, `tsa` and `acme` are optional modules.
    When the operator disables a module its routes are not registered and
    return a plain-text 404 from the HTTP router.
  license:
    name: See LICENSE in the repository
servers:
  - url: https://localhost:8080
    description: Default development listener; replace with the deployment's public base URL.
...
paths:
...
  /v1/crl/{file}:
    get:
      tags: [revocation]
      operationId: getCRL
      summary: DER-encoded CRL (CRL Distribution Point target)
      description: |
        `intermediate.crl` lists revoked leaf certificates; `root.crl` lists
        revoked certificates issued directly by the root.
      parameters:
        - name: file
          in: path
          required: true
          schema: {type: string, enum: [root.crl, intermediate.crl]}
      responses:
        "200":
          description: CRL.
          content:
            application/pkix-crl:
              schema: {type: string, format: binary}
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/tsa:
    post:
      tags: [tsa]
      operationId: timestamp
      summary: RFC 3161 time-stamp request
      description: SHA-1 message imprints are rejected; use SHA-256 or stronger.
      requestBody:
        required: true
        content:
          application/timestamp-query:
            schema: {type: string, format: binary, maxLength: 65536}
      responses:
        "200":
          description: DER-encoded TimeStampResp.
          content:
            application/timestamp-reply:
              schema: {type: string, format: binary}
        "400":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    mtls:
      type: mutualTLS
      description: |
        Client certificate issued by this CA (via `POST /v1/clients`, ACME, or
        the bootstrap admin certificate) with a role assigned and not revoked.
...
  responses:
    Error:
      description: RFC 9457 problem details. Branch on `type`; see the Problem schema for the catalogue.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457 problem details. `type` is a stable identifier; `detail` is
        human-readable and may change. Extension members give the data
        needed to correct the request (allowed values, required role, ...).
      required: [type, title, status]
      properties:
        type:
          type: string
          enum:
            - urn:trustmate:problem:malformed-json
            - urn:trustmate:problem:missing-field
            - urn:trustmate:problem:invalid-parameter
            - urn:trustmate:problem:unknown-profile
            - urn:trustmate:problem:invalid-csr
            - urn:trustmate:problem:invalid-role
            - urn:trustmate:problem:unknown-revocation-reason
            - urn:trustmate:problem:not-revocable
            - urn:trustmate:problem:already-revoked
            - urn:trustmate:problem:not-found
            - urn:trustmate:problem:profile-reload-failed
            - urn:trustmate:problem:profile-registry-unavailable
            - urn:trustmate:problem:unsupported-media-type
            - urn:trustmate:problem:body-unreadable
            - urn:trustmate:problem:body-too-large
            - urn:trustmate:problem:invalid-ocsp-request
            - urn:trustmate:problem:invalid-timestamp-request
            - urn:trustmate:problem:unsupported-timestamp-request
            - urn:trustmate:problem:client-certificate-required
            - urn:trustmate:problem:client-certificate-not-authorized
            - urn:trustmate:problem:client-certificate-revoked
            - urn:trustmate:problem:insufficient-role
            - urn:trustmate:problem:internal-error
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        available_profiles:
          type: array
          items: {type: string}
          description: On `unknown-profile`.
        missing_fields:
          type: array
          items: {type: string}
          description: On `missing-field`.
        allowed_values:
          type: array
          items: {type: string}
          description: On `invalid-role` and `unknown-revocation-reason`.
        required_role:
          $ref: "#/components/schemas/Role"
          description: On `insufficient-role` and `client-certificate-required`.
        role:
          $ref: "#/components/schemas/Role"
          description: The caller's role, on `insufficient-role`.
        client_serial:
          $ref: "#/components/schemas/Serial"
          description: The presented client certificate, on `client-certificate-not-authorized` and `client-certificate-revoked`.
        serial:
          $ref: "#/components/schemas/Serial"
          description: On `already-revoked`.
        parameter:
          type: string
          description: On `invalid-parameter`.
        maximum:
          type: integer
          description: Upper bound of the parameter, on `invalid-parameter`.
        expected_content_type:
          type: string
          description: On `unsupported-media-type`.
        max_bytes:
          type: integer
          description: On `body-too-large`.
        available:
          type: array
          items: {type: string}
          description: Valid CRL file names, on `not-found` from `/v1/crl/{file}`.
    Serial:
      type: string
      pattern: "^[0-9]+$"
      description: Certificate serial number in decimal.
    Role:
      type: string
      enum: [admin, manager]
      description: "`admin` is a superset of `manager`."
...